feat(platform-github): export test utilities for consumer testing
Add *Test modules for each service with mock layer factories:
- ActionRunnerTest.make() - mock inputs, outputs, logs, state
- ActionContextTest.make() - mock workflow context
- ActionClientTest.make() - mock Octokit responses
- ActionSummaryTest.make() - mock summary buffer

These are exported from the main package so consumers can
easily write tests for their actions:

  import { ActionRunner, ActionRunnerTest } from '@effect-native/platform-github'

  const test = ActionRunnerTest.make({ inputs: { name: 'world' } })
  const result = yield* ActionRunner.getInput('name').pipe(Effect.provide(test.layer))

Refactored internal tests to use the new modules as validation.
