+++ packages-native/platform-github/src/index.ts
@@ -17,7 +17,17 @@ export * as ActionRunner from "./ActionRunner.js"
 /**
  * @since 1.0.0
  */
+export * as ActionRunnerTest from "./ActionRunnerTest.js"
+
+/**
+ * @since 1.0.0
+ */
 export * as ActionContext from "./ActionContext.js"
+
+/**
+ * @since 1.0.0
+ */
+export * as ActionContextTest from "./ActionContextTest.js"
 
 /**
  * @since 1.0.0
@@ -27,7 +37,17 @@ export * as ActionClient from "./ActionClient.js"
 /**
  * @since 1.0.0
  */
+export * as ActionClientTest from "./ActionClientTest.js"
+
+/**
+ * @since 1.0.0
+ */
 export * as ActionSummary from "./ActionSummary.js"
+
+/**
+ * @since 1.0.0
+ */
+export * as ActionSummaryTest from "./ActionSummaryTest.js"
 
 /**
  * @since 1.0.0
