+++ packages-native/platform-github/test/ActionRunner.test.ts
@@ -1,169 +1,10 @@
 import { describe, expect, it } from "@effect/vitest"
 import * as Effect from "effect/Effect"
-import * as Layer from "effect/Layer"
-import { ActionInputError } from "../src/ActionError.js"
 import * as ActionRunner from "../src/ActionRunner.js"
+import * as ActionRunnerTest from "../src/ActionRunnerTest.js"
 
-// Test layer that mocks @actions/core behavior
-const makeTestLayer = (options: {
-  inputs?: Record<string, string>
-  outputs?: Record<string, unknown>
-  logs?: Array<{ level: string; message: string }>
-  groups?: string[]
-  secrets?: string[]
-  state?: Record<string, string>
-  env?: Record<string, string>
-  paths?: string[]
-  failed?: { message: string } | null
-}) => {
-  const inputs = options.inputs ?? {}
-  const outputs = options.outputs ?? {}
-  const logs: Array<{ level: string; message: string }> = options.logs ?? []
-  const groups: string[] = options.groups ?? []
-  const secrets: string[] = options.secrets ?? []
-  const state = options.state ?? {}
-  const env = options.env ?? {}
-  const paths: string[] = options.paths ?? []
-  let failed: { message: string } | null = options.failed ?? null
-
-  const runner: ActionRunner.ActionRunner = {
-    [ActionRunner.TypeId]: ActionRunner.TypeId,
-
-    getInput: (name, opts) =>
-      Effect.sync(() => {
-        const val = inputs[name] ?? ""
-        if (opts?.required && !val) {
-          throw new Error(`Input required and not supplied: ${name}`)
-        }
-        return opts?.trimWhitespace === false ? val : val.trim()
-      }),
-
-    getMultilineInput: (name, opts) =>
-      Effect.sync(() => {
-        const val = inputs[name] ?? ""
-        if (opts?.required && !val) {
-          throw new Error(`Input required and not supplied: ${name}`)
-        }
-        const lines = val.split("\n").filter((x) => x !== "")
-        return opts?.trimWhitespace === false ? lines : lines.map((l) => l.trim())
-      }),
-
-    getBooleanInput: (name, opts) =>
-      Effect.try({
-        try: () => {
-          const val = inputs[name] ?? ""
-          if (opts?.required && !val) {
-            throw new Error(`Input required and not supplied: ${name}`)
-          }
-          if (["true", "True", "TRUE"].includes(val)) return true
-          if (["false", "False", "FALSE"].includes(val)) return false
-          throw new TypeError(`Input does not meet YAML 1.2 "Core Schema" specification: ${name}`)
-        },
-        catch: (error) =>
-          new ActionInputError({
-            reason: "InvalidType",
-            name,
-            cause: error
-          })
-      }),
-
-    setOutput: (name, value) =>
-      Effect.sync(() => {
-        ;(outputs as Record<string, unknown>)[name] = value
-      }),
-
-    debug: (message) =>
-      Effect.sync(() => {
-        logs.push({ level: "debug", message })
-      }),
-
-    info: (message) =>
-      Effect.sync(() => {
-        logs.push({ level: "info", message })
-      }),
-
-    warning: (message) =>
-      Effect.sync(() => {
-        logs.push({ level: "warning", message: message instanceof Error ? message.message : message })
-      }),
-
-    error: (message) =>
-      Effect.sync(() => {
-        logs.push({ level: "error", message: message instanceof Error ? message.message : message })
-      }),
-
-    notice: (message) =>
-      Effect.sync(() => {
-        logs.push({ level: "notice", message: message instanceof Error ? message.message : message })
-      }),
-
-    startGroup: (name) =>
-      Effect.sync(() => {
-        groups.push(name)
-      }),
-
-    endGroup: () =>
-      Effect.sync(() => {
-        groups.pop()
-      }),
-
-    group: <A, E, R>(name: string, fn: () => Effect.Effect<A, E, R>) =>
-      Effect.acquireUseRelease(
-        Effect.sync(() => {
-          groups.push(name)
-        }),
-        () => fn(),
-        () =>
-          Effect.sync(() => {
-            groups.pop()
-          })
-      ),
-
-    exportVariable: (name, value) =>
-      Effect.sync(() => {
-        env[name] = value
-      }),
-
-    addPath: (path) =>
-      Effect.sync(() => {
-        paths.push(path)
-      }),
-
-    setSecret: (secret) =>
-      Effect.sync(() => {
-        secrets.push(secret)
-      }),
-
-    saveState: (name, value) =>
-      Effect.sync(() => {
-        state[name] = typeof value === "string" ? value : JSON.stringify(value)
-      }),
-
-    getState: (name) =>
-      Effect.sync(() => {
-        return state[name] ?? ""
-      }),
-
-    setFailed: (message) =>
-      Effect.sync(() => {
-        failed = { message: message instanceof Error ? message.message : message }
-      }),
-
-    getIDToken: () => Effect.succeed("mock-oidc-token")
-  }
-
-  return {
-    layer: Layer.succeed(ActionRunner.ActionRunner, runner),
-    outputs,
-    logs,
-    groups,
-    secrets,
-    state,
-    env,
-    paths,
-    getFailed: () => failed
-  }
-}
+// Re-export for backwards compatibility in tests
+const makeTestLayer = ActionRunnerTest.make
 
 describe("ActionRunner", () => {
   describe("getInput", () => {
