+++ packages-native/platform-github/test/ActionSummary.test.ts
@@ -1,142 +1,10 @@
 import { describe, expect, it } from "@effect/vitest"
 import * as Effect from "effect/Effect"
-import * as Layer from "effect/Layer"
-import { ActionSummaryError } from "../src/ActionError.js"
 import * as ActionSummary from "../src/ActionSummary.js"
+import * as ActionSummaryTest from "../src/ActionSummaryTest.js"
 
-// Test layer that mocks @actions/core.summary
-const makeTestLayer = (options?: {
-  writeError?: Error
-  clearError?: Error
-}) => {
-  let buffer = ""
-
-  const summary: ActionSummary.ActionSummary = {
-    [ActionSummary.TypeId]: ActionSummary.TypeId,
-
-    addRaw: (text, addEOL) => {
-      buffer += text
-      if (addEOL) buffer += "\n"
-      return summary
-    },
-
-    addEOL: () => {
-      buffer += "\n"
-      return summary
-    },
-
-    addHeading: (text, level = 1) => {
-      buffer += `<h${level}>${text}</h${level}>\n`
-      return summary
-    },
-
-    addCodeBlock: (code, lang) => {
-      const attrs = lang ? ` lang="${lang}"` : ""
-      buffer += `<pre${attrs}><code>${code}</code></pre>\n`
-      return summary
-    },
-
-    addList: (items, ordered = false) => {
-      const tag = ordered ? "ol" : "ul"
-      const listItems = items.map((item) => `<li>${item}</li>`).join("")
-      buffer += `<${tag}>${listItems}</${tag}>\n`
-      return summary
-    },
-
-    addTable: (rows) => {
-      const tableBody = rows
-        .map((row) => {
-          const cells = row
-            .map((cell) => {
-              const tag = cell.header ? "th" : "td"
-              const attrs: string[] = []
-              if (cell.colspan) attrs.push(`colspan="${cell.colspan}"`)
-              if (cell.rowspan) attrs.push(`rowspan="${cell.rowspan}"`)
-              const attrStr = attrs.length > 0 ? ` ${attrs.join(" ")}` : ""
-              return `<${tag}${attrStr}>${cell.data}</${tag}>`
-            })
-            .join("")
-          return `<tr>${cells}</tr>`
-        })
-        .join("")
-      buffer += `<table>${tableBody}</table>\n`
-      return summary
-    },
-
-    addDetails: (label, content) => {
-      buffer += `<details><summary>${label}</summary>${content}</details>\n`
-      return summary
-    },
-
-    addImage: (src, alt, opts) => {
-      const attrs = [`src="${src}"`, `alt="${alt}"`]
-      if (opts?.width) attrs.push(`width="${opts.width}"`)
-      if (opts?.height) attrs.push(`height="${opts.height}"`)
-      buffer += `<img ${attrs.join(" ")}>\n`
-      return summary
-    },
-
-    addLink: (text, href) => {
-      buffer += `<a href="${href}">${text}</a>\n`
-      return summary
-    },
-
-    addSeparator: () => {
-      buffer += "<hr>\n"
-      return summary
-    },
-
-    addBreak: () => {
-      buffer += "<br>\n"
-      return summary
-    },
-
-    addQuote: (text, cite) => {
-      const attrs = cite ? ` cite="${cite}"` : ""
-      buffer += `<blockquote${attrs}>${text}</blockquote>\n`
-      return summary
-    },
-
-    stringify: () => buffer,
-
-    isEmptyBuffer: () => buffer.length === 0,
-
-    emptyBuffer: () => {
-      buffer = ""
-      return summary
-    },
-
-    write: (_opts) =>
-      options?.writeError
-        ? Effect.fail(
-            new ActionSummaryError({
-              reason: "WriteFailed",
-              description: options.writeError.message,
-              cause: options.writeError
-            })
-          )
-        : Effect.succeed(summary),
-
-    clear: () =>
-      options?.clearError
-        ? Effect.fail(
-            new ActionSummaryError({
-              reason: "WriteFailed",
-              description: options.clearError.message,
-              cause: options.clearError
-            })
-          )
-        : Effect.sync(() => {
-            buffer = ""
-            return summary
-          })
-  }
-
-  return {
-    layer: Layer.succeed(ActionSummary.ActionSummary, summary),
-    getBuffer: () => buffer
-  }
-}
+// Re-export for backwards compatibility in tests
+const makeTestLayer = ActionSummaryTest.make
 
 describe("ActionSummary", () => {
   describe("content methods", () => {
